Setting up to build the gulp tasks: defining goals
