+++ www/markup/index.min.html
@@ -2,4 +2,4 @@
           (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
          })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
           ga('create', 'UA-92241125-1', 'auto');
-          ga('send', 'pageview');</script><!--========== For Web Crawlers and SEO =====--><meta charset="utf-8"><meta name="keywords" content="JCL,Certamen,Chancellors,Kyle,Michael,latin,Jeaprody,Stma"><meta name="author" content="Michael N., Kyle M."><meta name="description" content="Certamen Page for chancellorjcl.noip.me "></head><body><div id="page-container"></div><!--========== Run the React Scipt After it's container is ready(eventually relpace with async) ===--><script src="../javascript/index.min.js"></script></body></html>
+          ga('send', 'pageview');</script><!--========== For Web Crawlers and SEO =====--><meta charset="utf-8"><meta name="keywords" content="JCL,Certamen,Chancellors,Kyle,Michael,latin,Jeaprody,Stma"><meta name="author" content="Michael N., Kyle M."><meta name="description" content="Certamen Page for chancellorjcl.noip.me "></head><body><div id="page-container"></div><!--========== Run the React Scipt After it's container is ready(eventually relpace with async) ===--><script src="../javascript/index.min.js"></script><h1>test success</h1></body></html>
